modelserver/requirements.txt
absl-py==0.10.0
addict==2.2.1
albumentations==0.4.6
altair==4.1.0
argon2-cffi==20.1.0
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.6.0
bleach==3.1.5
blinker==1.4
boto3==1.14.57
botocore==1.17.57
cachetools==4.1.1
catalyst==20.9
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
constantly==15.1.0
cryptography==3.1
cycler==0.10.0
Cython==0.29.21
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
detectron2==0.2.1+cu101
Django==3.1
docutils==0.15.2
easydict==1.9
entrypoints==0.3
enum-compat==0.0.3
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
fvcore==0.1.1.post20200716
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.8
glob2==0.7
google-api-core==1.22.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-cloud-speech==2.0.0
googleapis-common-protos==1.52.0
graphviz==0.8.4
grpcio==1.31.0
hyperlink==20.0.1
idna==2.10
iglovikov-helper-functions==0.0.41
imagecorruptions==1.1.0
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
incremental==17.5.0
insightface==0.1.5
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.4.2
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jpeg4py==0.1.4
JPype1==1.0.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
konlpy==0.5.2
lazy-object-proxy==1.4.3
libarchive-c==2.9
libcst==0.3.10
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
mxnet-cu101==1.7.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.4
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.42
opencv-python-headless==4.4.0.44
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
plotly==4.9.0
portalocker==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
proto-plus==1.10.0
protobuf==3.13.0
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycosat==0.6.3
pycparser==2.20
pydeck==0.5.0b1
pydot==1.4.1
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.23.0
requests-oauthlib==1.3.0
retinaface-pytorch==0.0.7
retrying==1.3.3
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
service-identity==18.1.0
Shapely==1.7.1
six==1.15.0
smart-open==2.2.1
smmap==3.0.4
soupsieve==2.0.1
sqlparse==0.3.1
streamlit==0.66.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.1
toolz==0.10.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
tweepy==3.9.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
urllib3==1.25.8
validators==0.18.1
visual-attention==0.0.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
zipp==3.1.0
zope.interface==5.1.0
